website clone
This is a website clone experiment of jumia shopping website. credits and rights for all codes written in html,JavaScript,css are all jumia's. the aim of this experiment is simply to test my website cloning abilities and study the behaviour of cloned websites which i consider an important aspect in my programming journey.